pymityma/requirements.txt
aiofiles==0.7.0
fastapi==0.67.0
flake8==3.9.2
isort==5.9.3
pip-chill==1.0.1
psycopg2-binary==2.9.3
pytesseract==0.3.4
pytest==6.2.4
python-multipart==0.0.5
requests==2.26.0
flask==2.1.2
flask-cors==3.0.10
sqlalchemy==1.4.40
python-dotenv==0.21.0
flask-classful==0.14.2
werkzeug<=2.1.2   
requests-pkcs12==1.13
flask-marshmallow==0.14.0
flask-sqlalchemy==2.5.1
marshmallow-sqlalchemy==0.28.1
flask-migrate==3.1.0 
gunicorn==20.1.0
python-dateutil~=2.9.0.post0
alembic~=1.14.0